+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unversioned
 
+* Automatically configure Rails cache for API apps
 * Add config defaults that are common across our initializers
 
 # 15.0.1